FRCEvent made and integrated
Some stuff to think about;
I don't think FRCEvent should do any output at all if possible; only
collect and process data (and have other things do the outputting).
In that spirit, no longer is it writing out to the table the little
messages. Instead we'll have Main.js do that.
(This is good so that we can have multiple FRCEvents and not have them
messing with one anotheR).
.status represents the status message for that FRC Event.