Removed excanvas
Dropping support for IE<9 because of the many idiosyncracies found
there. <-- I don't know how to spell this word.

IE8,7,6, among other things, does not allow the modification of tables
rendering on the page already through the simple methods that I use.
This is avoidable but unnecessary as these browsers represent now a
shrinking fraction of users, probably especially for the tech savy who
at the least upgrade to the modern IE9,10.